// src/shared/model/top-home-slides.model.ts
export interface ITopHomeSlides {
    id?: any;
    titleAr?: string;
    titleFr?: string;
    titleEn?: string;
    subTitleAr?: string;
    subTitleFr?: string;
    subTitleEn?: string;
    descriptionAr?: string;
    descriptionFr?: string;
    descriptionEn?: string;
}
export const defaultValue: Readonly<ITopHomeSlides> = {};